refactor(nuxt-app): move base mapping to nuxt-ripple and fix imports
