+++ packages/nuxt-ripple/mapping/base/sidebar-contacts/__test__/rawdata.ts
