+++ packages/nuxt-ripple/mapping/base/sidebar-contacts/__test__/sidebar-contacts-mapping.test.ts
