+++ packages/nuxt-ripple/mapping/base/sidebar-contacts/sidebar-contacts-mapping-types.ts
