+++ packages/nuxt-ripple/mapping/base/sidebar-social-share/__test__/rawdata.ts
