+++ packages/nuxt-ripple/mapping/base/sidebar-social-share/__test__/sidebar-social-share-mapping.test.ts
