+++ packages/nuxt-ripple/mapping/base/sidebar-social-share/sidebar-social-share-mapping.ts
