+++ packages/nuxt-ripple/mapping/base/sidebar-whats-next/__test__/sidebar-whats-next-mapping.test.ts
