+++ packages/nuxt-ripple/mapping/index.ts
@@ -1,0 +1,6 @@
+export {
+  default as base,
+  tidePageBaseIncludes,
+  tidePageBaseMapping
+} from './base'
+export { default as site } from './site'
