+++ packages/nuxt-ripple/mapping/site/alerts/__test__/site-alerts-mapping.test.ts
