+++ packages/nuxt-ripple/mapping/site/alerts/site-alerts-mapping.ts
