+++ packages/nuxt-ripple/server/plugins/site.ts
@@ -10,4 +10,3 @@ function defineNitroPlugin(def: NitroAppPlugin): NitroAppPlugin {
 export default defineNitroPlugin((NitroApp) => {
   NitroApp.tide.siteApi.setSiteMapping(siteMapping)
 })
-11
