+++ packages/ripple-tide-api/src/index.ts
@@ -4,10 +4,6 @@ export { default as TidePageApi } from './services/tide-page.js'
 export { default as TideSiteApi } from './services/tide-site.js'
 export { default as logger } from './logger/logger.js'
 export { default as createHandler } from './utils/createHandler.js'
-export {
-  tidePageBaseMapping,
-  tidePageBaseIncludes
-} from './mapping/tide-page-base-mapping.js'
 export * from './utils/define-module.js'
 export * from './utils/formatPriceRange.js'
 export {
