+++ packages/ripple-tide-event/mapping/index.ts
@@ -1,21 +1,15 @@
 import {
-  // getAddress,
-  drupalField,
+  getAddress,
   getField,
   getBodyFromField,
   getLinkFromField,
-  formatPriceRange,
+  formatPriceRange
+} from '@dpc-sdp/ripple-tide-api'
+import {
   tidePageBaseMapping,
   tidePageBaseIncludes
-} from '@dpc-sdp/ripple-tide-api'
+} from '@dpc-sdp/nuxt-ripple/mapping'
 import type { IRplTideModuleMapping } from '@dpc-sdp/ripple-tide-api/types'
-
-const getAddress = (field: drupalField) => `
- ${field.address_line1}${field.address_line1 ? ', ' : ''}
- ${field.address_line2}${field.address_line2 ? ', ' : ''}
- ${field.locality}${field.address_line2 || field.locality ? ', ' : ''}
- ${field.administrative_area} ${field.postal_code}
-`
 
 const tideEventModule: IRplTideModuleMapping = {
   mapping: {
