+++ packages/ripple-tide-grant/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@dpc-sdp/ripple-tide-api": "workspace:*",
-    "@dpc-sdp/ripple-ui-core": "workspace:*"
+    "@dpc-sdp/ripple-ui-core": "workspace:*",
+    "@dpc-sdp/nuxt-ripple": "workspace:*"
   }
 }
