+++ packages/ripple-tide-news/mapping/index.ts
@@ -1,11 +1,13 @@
 import type { IRplTideModuleMapping } from '@dpc-sdp/ripple-tide-api/types'
 import {
-  tidePageBaseMapping,
-  tidePageBaseIncludes,
   getImageFromField,
   getBodyFromField,
   getField
 } from '@dpc-sdp/ripple-tide-api'
+import {
+  tidePageBaseMapping,
+  tidePageBaseIncludes
+} from '@dpc-sdp/nuxt-ripple/mapping'
 
 const tideNewsModule: IRplTideModuleMapping = {
   mapping: {
