+++ packages/ripple-tide-publication/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@dpc-sdp/ripple-tide-api": "workspace:*",
     "@dpc-sdp/ripple-tide-landing-page": "workspace:*",
-    "@dpc-sdp/ripple-ui-core": "workspace:*"
+    "@dpc-sdp/ripple-ui-core": "workspace:*",
+    "@dpc-sdp/nuxt-ripple": "workspace:*"
   }
 }
